Move testsutils from external-plugins to pkg dir Update new testutils path.
