+++ external-plugins/rehearse/plugin/handler/handler_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/sirupsen/logrus"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/util/sets"
-	"kubevirt.io/project-infra/external-plugins/testutils"
+	"kubevirt.io/project-infra/pkg/testutils"
 	prowapi "sigs.k8s.io/prow/pkg/apis/prowjobs/v1"
 	"sigs.k8s.io/prow/pkg/config"
 	"sigs.k8s.io/prow/pkg/git/localgit"
